sst start: fix --outputs-file option not working
