+++ packages/cli/scripts/deploy.js
@@ -4,7 +4,7 @@ const path = require("path");
 const fs = require("fs-extra");
 
 const paths = require("./util/paths");
-const { synth, deploy } = require("./util/cdkHelpers");
+const { synth, deploy, writeOutputsFile } = require("./util/cdkHelpers");
 const { STACK_DEPLOY_STATUS } = require("@serverless-stack/core");
 
 module.exports = async function (argv, config, cliInfo) {
@@ -23,10 +23,7 @@ module.exports = async function (argv, config, cliInfo) {
   // Run CDK Deploy
   const stacksData = await deploy(cliInfo.cdkOptions, stackName);
 
-  // This is native CDK option. According to CDK documentation:
-  // If an outputs file has been specified, create the file path and write stack outputs to it once.
-  // Outputs are written after all stacks have been deployed. If a stack deployment fails,
-  // all of the outputs from successfully deployed stacks before the failure will still be written.
+  // Write outputsFile
   if (argv.outputsFile) {
     await writeOutputsFile(
       stacksData,
@@ -41,18 +38,3 @@ module.exports = async function (argv, config, cliInfo) {
 
   return stacksData;
 };
-
-async function writeOutputsFile(stacksData, outputsFileWithPath) {
-  const stackOutputs = stacksData.reduce((acc, { name, outputs }) => {
-    if (Object.keys(outputs || {}).length > 0) {
-      return { ...acc, [name]: outputs };
-    }
-    return acc;
-  }, {});
-
-  fs.ensureFileSync(outputsFileWithPath);
-  await fs.writeJson(outputsFileWithPath, stackOutputs, {
-    spaces: 2,
-    encoding: "utf8",
-  });
-}
